+++ hydra-tui/src/Hydra/TUI/Drawing.hs
@@ -1,6 +1,4 @@
 {-# LANGUAGE DuplicateRecordFields #-}
-{-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE TypeApplications #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Hydra.TUI.Drawing where
@@ -8,33 +6,22 @@ module Hydra.TUI.Drawing where
 import Hydra.Prelude hiding (Down, State)
 
 import Brick
-import Hydra.Cardano.Api hiding (Active)
-
-import Brick.Forms (
-  renderForm,
- )
-import Brick.Widgets.Border (borderWithLabel, hBorder, hBorderWithLabel)
+import Brick.Widgets.Border (borderWithLabel)
 import Brick.Widgets.Border.Style (unicodeRounded)
-import Brick.Widgets.List qualified as BrickList
-import Cardano.Api.UTxO qualified as UTxO
-import Data.Map qualified as Map
 import Data.Text qualified as T
-import Data.Time (defaultTimeLocale, formatTime, utctDayTime)
-import Data.Time.Format (FormatTime)
-import Data.Time.LocalTime (TimeZone, utcToLocalTime)
 import Graphics.Vty qualified as Vty
+import Hydra.Cardano.Api hiding (Active)
 import Hydra.Chain.CardanoClient (CardanoClient (..))
-import Hydra.Chain.Direct.State ()
 import Hydra.Client (Client (..))
-import Hydra.Network (Host)
-import Hydra.TUI.Drawing.Utils (drawHex, drawShow, prettyHeadId)
-import Hydra.TUI.Logging.Types (LogMessage (..), Severity (..), logMessagesL)
+import Hydra.TUI.Drawing.EventHistoryTab (drawEventHistoryTab)
+import Hydra.TUI.Drawing.FundsTab (drawFocusPanel, drawFundsTab)
+import Hydra.TUI.Drawing.MainTab (drawMainTab)
 import Hydra.TUI.Model
 import Hydra.TUI.Style hiding (style)
-import Hydra.Tx (IsTx (..), Party (..))
 import Lens.Micro ((^.), (^?))
 
--- | Main draw function
+-- | Root draw function wiring together the tab bar, active tab, status bar,
+-- and action bar.
 draw :: CardanoClient -> Client Tx IO -> RootState -> [Widget Name]
 draw cardanoClient hydraClient s =
   pure $
@@ -110,184 +97,6 @@ drawModalTab CardanoClient{networkId} Client{sk} s =
         Disconnected -> emptyWidget
         Connected k -> drawFocusPanel networkId (getVerificationKey sk) (s ^. nowL) k
 
-drawMainTab :: CardanoClient -> Client Tx IO -> RootState -> Widget Name
-drawMainTab CardanoClient{networkId} Client{sk} s =
-  borderWithLabel panelLabel $
-    vBox
-      [ -- Status strip: plain text, no border characters so junctions don't misalign
-        padLeftRight 1 $
-          hBox . intersperse (txt "   ") $
-            [ drawConnectedStatus s
-            , drawNetworkState connState
-            , drawChainSyncedState connState
-            ]
-      , padLeftRight 1 $
-          hBox
-            [ hLimit 32 $
-                vBox
-                  [ borderWithLabel (withAttr neutral $ txt " Peers ") $
-                      padLeftRight 1 drawPeersSection
-                  , borderWithLabel (withAttr neutral $ txt " Head parties ") $
-                      padLeftRight 1 drawHeadPartiesSection
-                  ]
-            , padLeft (Pad 1) $
-                borderWithLabel (withAttr neutral $ txt " Head state ") $
-                  padLeftRight 1 drawHeadStateSection
-            ]
-      , hBorderWithLabel (withAttr neutral $ txt " Recent events ")
-      , vLimitPercent 50 $
-          viewport "recent-events" Vertical $
-            vBox $
-              map (drawEventListItem (s ^. timeZoneL) False) (take 20 (s ^. logStateL . logMessagesL))
-      ]
- where
-  connState = s ^. connectedStateL
-  ownAddress = mkVkAddress networkId (getVerificationKey sk)
-
-  panelLabel :: Widget Name
-  panelLabel = case connState of
-    Disconnected -> withAttr neutral $ txt " Status "
-    Connected (Connection{headState}) -> case headState of
-      Idle -> withAttr neutral $ txt " Idle "
-      Active (ActiveLink{headId, activeHeadState}) ->
-        let stateStr = case activeHeadState of
-              Open{} -> "Open"
-              Closed{} -> "Closed"
-              FanoutPossible -> "Ready to Fanout"
-              Final -> "Finalized"
-            hid = prettyHeadId headId
-         in withAttr neutral (txt " ")
-              <+> withAttr headStateA (txt stateStr)
-              <+> withAttr neutral (txt (" · " <> hid <> " "))
-
-  drawPeersSection :: Widget Name
-  drawPeersSection = case connState of
-    Disconnected -> withAttr neutral $ txt "Peers"
-    Connected c -> drawPeers connState (c ^. peersL)
-
-  drawHeadPartiesSection :: Widget Name
-  drawHeadPartiesSection = case connState of
-    Disconnected -> emptyWidget
-    Connected c -> case c ^. headStateL of
-      Idle -> withAttr neutral $ txt "none"
-      Active (ActiveLink{parties}) ->
-        case c ^. meL of
-          Unidentified -> vBox $ map (drawParty mempty) parties
-          Identified ownParty -> drawPartiesWithOwnHighlighted ownParty parties
-
-  drawHeadStateSection :: Widget Name
-  drawHeadStateSection = case connState of
-    Disconnected -> withAttr neutral $ txt "Not connected."
-    Connected c -> case c ^. headStateL of
-      Idle -> withAttr neutral $ txt "Head is idle."
-      Active (ActiveLink{utxo, pendingIncrements, pendingUTxOToDecommit, activeHeadState}) ->
-        case activeHeadState of
-          Open{} ->
-            vBox $
-              [ withAttr neutral $ txt "UTxO"
-              , drawUTxO (highlightOwnAddress ownAddress) utxo
-              ]
-                <> [ withAttr infoA $ txt $ "  ↑ " <> show (length pendingIncrements) <> " pending commit(s)"
-                   | not (null pendingIncrements)
-                   ]
-                <> [ withAttr infoA $ txt "  ↓ pending decommit"
-                   | pendingUTxOToDecommit /= mempty
-                   ]
-          Closed (ClosedState{contestationDeadline}) ->
-            drawRemainingContestationPeriod contestationDeadline (s ^. nowL)
-          FanoutPossible ->
-            withAttr positive $ txt "Contestation period passed — ready to fan out."
-          Final ->
-            vBox
-              [ withAttr positive $ txt "Head finalized."
-              , drawUTxO (highlightOwnAddress ownAddress) utxo
-              ]
-
-drawFundsTab :: CardanoClient -> Client Tx IO -> RootState -> Widget Name
-drawFundsTab CardanoClient{networkId} Client{sk} s =
-  borderWithLabel (withAttr neutral $ txt " Funds ") $
-    vBox
-      [ hBorderWithLabel (withAttr neutral $ txt " L2 State ")
-      , vLimitPercent 50 $ padLeftRight 1 drawL2
-      , hBorderWithLabel (withAttr neutral $ txt " L1 Wallet ")
-      , padLeftRight 1 $ drawL1WalletPanel (s ^. l1UTxOL) ownAddress (s ^. nowL)
-      ]
- where
-  vk = getVerificationKey sk
-  ownAddress = mkVkAddress networkId vk
-  drawL2 = case s ^. connectedStateL of
-    Disconnected -> withAttr neutral $ txt "Not connected."
-    Connected k -> drawFocusPanel networkId vk (s ^. nowL) k
-
-drawL1WalletPanel :: Maybe (Map TxIn (TxOut CtxUTxO)) -> AddressInEra -> UTCTime -> Widget Name
-drawL1WalletPanel Nothing _ now =
-  withAttr neutral $ padAll 1 $ txt (spinnerFrame now <> " Refreshing…")
-drawL1WalletPanel (Just utxo) ownAddress _
-  | Map.null utxo = withAttr neutral $ padAll 1 $ txt "No UTxO found."
-  | otherwise =
-      vBox
-        [ withAttr neutral $ txt "Wallet UTxO"
-        , drawUTxO (highlightOwnAddress ownAddress) (UTxO.fromMap utxo)
-        ]
-
-spinnerFrame :: UTCTime -> Text
-spinnerFrame now =
-  let frames = ["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"] :: [Text]
-      s = floor (realToFrac (utctDayTime now) :: Double) `mod` (10 :: Int)
-   in case drop s frames of
-        (f : _) -> f
-        [] -> "⠋"
-
-drawEventHistoryTab :: RootState -> Widget Name
-drawEventHistoryTab s =
-  borderWithLabel (withAttr neutral $ txt " Event History ") $
-    vBox
-      [ vLimitPercent 50 $ BrickList.renderList (drawEventListItem tz) True (s ^. eventHistoryListL)
-      , hBorderWithLabel (withAttr neutral $ txt detailLabel)
-      , viewport "event-detail" Vertical $
-          padLeftRight 1 $
-            drawEventDetail tz rawView (BrickList.listSelectedElement (s ^. eventHistoryListL))
-      ]
- where
-  tz = s ^. timeZoneL
-  rawView = s ^. eventDetailRawL
-  detailLabel = if rawView then " Detail (raw)  d:summary " else " Detail  d:raw "
-
-drawEventListItem :: TimeZone -> Bool -> LogMessage -> Widget Name
-drawEventListItem tz selected (LogMessage{message, severity, time}) =
-  let ts = formatTime defaultTimeLocale "%b %d %H:%M:%S" (utcToLocalTime tz time)
-      line = str ts <+> txt "  " <+> txt severityIcon <+> txt "  " <+> txt message
-      styled = case severity of
-        Success | not selected -> withAttr infoA line
-        _ -> line
-   in if selected then withAttr BrickList.listSelectedAttr styled else styled
- where
-  severityIcon = case severity of
-    Success -> "✓"
-    Info -> "·"
-    Error -> "✗"
-
-drawEventDetail :: TimeZone -> Bool -> Maybe (Int, LogMessage) -> Widget Name
-drawEventDetail _ _ Nothing = withAttr neutral $ txt "No event selected."
-drawEventDetail tz rawView (Just (_, LogMessage{detail, severity, time, rawJson})) =
-  vBox
-    [ hBox
-        [ withAttr (severityToAttr severity) $ txt (severityLabel severity)
-        , txt "  "
-        , str $ formatTime defaultTimeLocale "%b %d %Y  %H:%M:%S" (utcToLocalTime tz time)
-        ]
-    , txt " "
-    , body
-    ]
- where
-  severityLabel = \case
-    Success -> "Success"
-    Info -> "Info"
-    Error -> "Error"
-  body
-    | rawView = txtWrap rawJson
-    | otherwise = txtWrap detail
-
 drawActionBar :: RootState -> Widget n
 drawActionBar s =
   hBox . intersperse (txt "  ") $ map drawAction actions
@@ -321,173 +130,3 @@ drawActionBar s =
 
   drawAction :: (Text, Text) -> Widget n
   drawAction (key, rest) = withAttr keyA (txt key) <+> modifyDefAttr (`Vty.withStyle` Vty.italic) (txt rest)
-
-drawRemainingDepositDeadline :: UTCTime -> UTCTime -> Widget Name
-drawRemainingDepositDeadline deadline now =
-  let remaining = diffUTCTime deadline now
-   in if remaining > 0
-        then txt "Remaining time to deposit: " <+> str (renderTime remaining)
-        else withAttr negative $ txt "Deposit deadline passed, ready to recover."
-
-drawPendingIncrement :: AddressInEra -> [PendingIncrement] -> UTCTime -> Widget Name
-drawPendingIncrement ownAddress pendingIncrements now =
-  vBox $ foldl' pendingWidget [] pendingIncrements
- where
-  pendingWidget acc = \case
-    PendingIncrement{utxoToCommit, deposit, depositDeadline, status} ->
-      acc
-        <> [ txt ("id: " <> show deposit)
-           , txt ("status: " <> show status)
-           , drawUTxO (highlightOwnAddress ownAddress) utxoToCommit
-           , drawRemainingDepositDeadline depositDeadline now
-           , hBorder
-           ]
-
-drawFocusPanelOpen :: NetworkId -> VerificationKey PaymentKey -> UTxO -> UTxO -> [PendingIncrement] -> UTCTime -> OpenScreen -> Widget Name
-drawFocusPanelOpen networkId vk utxo pendingUTxOToDecommit pendingIncrements now = \case
-  OpenHome ->
-    vBox
-      [ withAttr neutral (txt "Active UTxO")
-      , drawUTxO (highlightOwnAddress ownAddress) utxo
-      , hBorder
-      , withAttr neutral (txt "Pending decommit")
-      , drawUTxO (highlightOwnAddress ownAddress) pendingUTxOToDecommit
-      , hBorder
-      , withAttr neutral (txt "Pending commits")
-      , drawPendingIncrement ownAddress pendingIncrements now
-      ]
-  LoadingUTxOForIncrement ->
-    padAll 1 $ withAttr neutral $ txt "Querying Cardano node for available UTxO…"
-  SelectingUTxO x -> renderForm x
-  SelectingUTxOToDecommit x -> renderForm x
-  SelectingUTxOToIncrement x -> renderForm x
-  SelectingDepositIdToRecover x -> renderForm x
-  EnteringAmount _ x -> renderForm x
-  SelectingRecipient _ _ x -> renderForm x
-  EnteringRecipientAddress _ _ x -> renderForm x
-  ConfirmingClose x -> vBox [txt "Confirm close:", renderForm x]
- where
-  ownAddress = mkVkAddress networkId vk
-
-drawFocusPanelClosed :: UTCTime -> ClosedState -> Widget Name
-drawFocusPanelClosed now (ClosedState{contestationDeadline}) = drawRemainingContestationPeriod contestationDeadline now
-
-drawFocusPanelFinal :: NetworkId -> VerificationKey PaymentKey -> UTxO -> Widget Name
-drawFocusPanelFinal networkId vk utxo =
-  vBox
-    [ txt ("Distributed UTxO — total: " <> renderAda (balance @Tx utxo))
-    , padLeft (Pad 2) (drawUTxO (highlightOwnAddress ownAddress) utxo)
-    ]
- where
-  ownAddress = mkVkAddress networkId vk
-
-highlightOwnAddress :: AddressInEra -> AddressInEra -> Widget n
-highlightOwnAddress ownAddress a =
-  withAttr (if a == ownAddress then own else mempty) $ drawAddress a
-
-drawFocusPanel :: NetworkId -> VerificationKey PaymentKey -> UTCTime -> Connection -> Widget Name
-drawFocusPanel networkId vk now (Connection{headState}) = case headState of
-  Idle -> withAttr neutral $ txt "Head is idle."
-  Active (ActiveLink{utxo, pendingUTxOToDecommit, pendingIncrements, activeHeadState}) -> case activeHeadState of
-    Open x -> drawFocusPanelOpen networkId vk utxo pendingUTxOToDecommit pendingIncrements now x
-    Closed x -> drawFocusPanelClosed now x
-    FanoutPossible -> txt "Ready to fanout!"
-    Final -> drawFocusPanelFinal networkId vk utxo
-
-drawRemainingContestationPeriod :: UTCTime -> UTCTime -> Widget Name
-drawRemainingContestationPeriod deadline now =
-  let remaining = diffUTCTime deadline now
-   in if remaining > 0
-        then txt "Remaining time to contest: " <+> str (renderTime remaining)
-        else withAttr positive $ txt "Contestation period passed — ready to fan out."
-
-drawPartiesWithOwnHighlighted :: Party -> [Party] -> Widget n
-drawPartiesWithOwnHighlighted k = drawParties (\p -> drawParty (if k == p then own else mempty) p)
-
-drawParties :: (Party -> Widget n) -> [Party] -> Widget n
-drawParties f xs = vBox $ map f xs
-
-drawNetworkState :: ConnectedState -> Widget n
-drawNetworkState s =
-  txt "Network: " <+> case s of
-    Disconnected{} -> withAttr negative $ txt "Unknown"
-    Connected{connection = Connection{networkState}} ->
-      case networkState of
-        Nothing -> withAttr negative $ txt "Unknown"
-        Just NetworkConnected -> withAttr positive $ txt "Connected"
-        Just NetworkDisconnected -> withAttr negative $ txt "Disconnected"
-
-drawChainSyncedState :: ConnectedState -> Widget n
-drawChainSyncedState s =
-  txt "Chain: " <+> case s of
-    Disconnected{} -> withAttr negative $ txt "Unknown"
-    Connected{connection = Connection{chainSyncedStatus}} ->
-      case chainSyncedStatus of
-        InSync -> withAttr positive $ txt "In sync"
-        CatchingUp -> withAttr negative $ txt "Catching up"
-
-drawPeers :: ConnectedState -> [(Host, PeerStatus)] -> Widget n
-drawPeers s peers = vBox rest
- where
-  rest = case s of
-    Connected{connection = Connection{networkState = Just NetworkConnected}} ->
-      if null peers
-        then [withAttr neutral $ txt "none"]
-        else map drawPeer peers
-    _ -> [withAttr neutral $ txt "unknown"]
-
-  drawPeer :: (Host, PeerStatus) -> Widget n
-  drawPeer (host, status) =
-    withAttr (statusAttr status) (txt "●") <+> txt " " <+> drawShow host
-
-  statusAttr = \case
-    PeerIsConnected -> infoA
-    PeerIsDisconnected -> negative
-    PeerIsUnknown -> neutral
-
-drawAddress :: AddressInEra -> Widget n
-drawAddress addr = txt (serialiseAddress addr)
-
-drawConnectedStatus :: RootState -> Widget n
-drawConnectedStatus RootState{nodeHost, connectedState} =
-  txt "API: " <+> case connectedState of
-    Disconnected -> withAttr negative $ str $ "Connecting to " <> show nodeHost
-    Connected _ -> withAttr positive $ txt "Connected"
-
-drawParty :: AttrName -> Party -> Widget n
-drawParty x Party{vkey} = withAttr x $ drawHex vkey
-
-renderTime :: (Ord t, Num t, FormatTime t) => t -> String
-renderTime r
-  | r < 0 = "-" <> renderTime (negate r)
-  | otherwise = formatTime defaultTimeLocale "%dd %Hh %Mm %Ss" r
-
--- | Render a lovelace value as ADA with the ₳ symbol.
-renderAda :: Value -> Text
-renderAda v =
-  let Coin l = selectLovelace v
-      (ada, frac) = abs l `divMod` 1_000_000
-      fracStr = show frac
-      padded = replicate (6 - length fracStr) '0' <> fracStr
-      sign = if l < 0 then "-" else ""
-   in sign <> "₳ " <> show ada <> "." <> toText padded
-
--- | Render a single UTxO entry showing the TxIn and its ADA value (with blue ADA amount).
-drawUTxOEntryAda :: (TxIn, TxOut CtxUTxO) -> Widget n
-drawUTxOEntryAda (txin, TxOut _ val _ _) =
-  txt (T.drop 54 (renderTxIn txin) <> "  ") <+> withAttr infoA (txt (renderAda val))
-
-drawUTxO :: (AddressInEra -> Widget n) -> UTxO -> Widget n
-drawUTxO f utxo =
-  let byAddress =
-        Map.foldrWithKey
-          (\k v@TxOut{txOutAddress = addr} -> Map.unionWith (++) (Map.singleton addr [(k, v)]))
-          mempty
-          $ UTxO.toMap utxo
-   in vBox
-        [ vBox
-          [ f addr
-          , padLeft (Pad 2) $ vBox (drawUTxOEntryAda <$> u)
-          ]
-        | (addr, u) <- Map.toList byAddress
-        ]
